requirements.txt
pandas
SQLAlchemy
pytest
numpy
pyarrow
pyyaml
openpyxl
psycopg2
logger
openmeteo-requests
requests-cache
retry-requests
gcsfs